feat: Change Activity Time to 5 minutes
Move Ready funciotns to outher file
